requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib
boto3
python-dotenv
python-multipart
